pyproject.toml
[project]
name = "app-revision-quizz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "google-genai>=1.49.0",
    "google-generativeai>=0.8.5",
    "markitdown>=0.0.2",
    "openai>=2.7.1",
    "psycopg[binary]>=3.2.12",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
